Replace `Get<float>` by `Get<double>` in property parsing
With latest version of Boost (1.88), `Get<float>` raises "RuntimeError: bad
lexical cast: source type value could not be interpreted as target"
exception for floating point values that don't exactly fit `float` precision.

Such values can be very innocent-looking: for instance, 39.5978 cannot be
stored as is in a float, but will be stored as 39.597801, what will make
Boost raise, and give the impression the problem is random.

To fix it, the only solution was to parse floating point values as
double and then narrow it to float.

+++ src/luxrays/accelerators/bvhaccel.cpp
@@ -53,7 +53,7 @@ BVHParams BVHAccel::ToBVHParams(const Properties &props) {
 	const int costSamples = props.Get(Property("accelerator.bvh.costsamples")(0)).Get<int>();
 	const int isectCost = props.Get(Property("accelerator.bvh.isectcost")(80)).Get<int>();
 	const int travCost = props.Get(Property("accelerator.bvh.travcost")(10)).Get<int>();
-	const float emptyBonus = props.Get(Property("accelerator.bvh.emptybonus")(.5)).Get<float>();
+	const float emptyBonus = props.Get(Property("accelerator.bvh.emptybonus")(.5)).Get<double>();
 	
 	BVHParams params;
 	// Make sure treeType is 2, 4 or 8
